requirements.txt
faiss_cpu==1.7.4
langchain==0.0.250
python-dotenv==1.0.0
textract==1.6.5
transformers==4.29.2
